% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_counter.R
\name{na_counter}
\alias{na_counter}
\title{Summarize the missing values that exist in a dataframe}
\usage{
na_counter(X, col_index = c())
}
\arguments{
\item{X}{tbl_df, df, data.frame}

\item{col_index}{vector of columns}
}
\value{
tbl_df
}
\description{
Summarize the missing values that exist in a dataframe
}
\examples{

df <- tibble::tibble(feat1 = c(1:3, NA, 5:7, NA, NA, NA))

na_counter(X=df, col_index = c("feat1"))

}
