% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdizer.R
\name{stdizer}
\alias{stdizer}
\title{Standardize features within a dataframe.}
\usage{
stdizer(X, col_index = NULL, method = "mean_sd", method_args)
}
\arguments{
\item{X}{tbl_df, df, data.frame}

\item{col_index}{vector of strings or integers}

\item{method}{string}

\item{method_args}{vector}
}
\value{
tbl_df
}
\description{
Accepts a tbl_df, df, or data.frame and returns a tbl_df with standardized features
}
\examples{
df <- tibble::tibble(feat1 = c(1:100), feat2 = runif(100, min = 0, max = 5000))

df_stdized <- stdizer(X = df, col_index = c(1, 2), method = 'mean_sd')
}
