% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{splitter}
\alias{splitter}
\title{Split data into train and test sets.}
\usage{
splitter(X, target, split_size, seed)
}
\arguments{
\item{X}{tbl_df, tbl, data.frame}

\item{target}{string}

\item{split_size}{numeric}

\item{seed}{numeric}
}
\value{
tbl_df
}
\description{
Takes in a tbl_df, df, or data.frame and returns a tbl_df for each of X_train, X_test, y_train, y_test.
}
\examples{
require(tibble)
df <- tibble::tibble(X = c(1:10), y = rbinom(10, 1, 0.5))

split_data <- splitter(df, target = "y", split_size = 0.8, seed=42)
}
